requirements.txt
langchain-core
langchain
langchain-google-genai
google-generativeai
langchain-huggingface
transformers
huggingface-hub
langchain-community
langchain-text-splitters
python-dotenv
langchain_chroma
flask
werkzeug.security
markdown